sha1_file: add repository argument to check_and_freshen_nonlocal
Add a repository argument to allow the check_and_freshen_nonlocal
caller to be more specific about which repository to act on. This is a
small mechanical change; it doesn't change the implementation to
handle repositories other than the_repository yet.

As with the previous commits, use a macro to catch callers passing a
repository other than the_repository at compile time.

Signed-off-by: Stefan Beller <sbeller@google.com>
Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>