object-store: add repository argument to has_loose_object_nonlocal
Add a repository argument to allow the caller of
has_loose_object_nonlocal to be more specific about which repository
to handle. This is a small mechanical change; it doesn't change the
implementation to handle repositories other than the_repository yet.

As with the previous commits, use a macro to catch callers passing a
repository other than the_repository at compile time.

Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>