repository: keep track of all open repositories
When Git mmap's packfiles into memory it has to keep track of the
memory management itself to not use too much memory. Each repository
has its own set of packfiles that it keeps track of. To scan all the
packfiles that are currently mmap'ed, we need to keep track of all
repositories that are currently opened.

Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>
Signed-off-by: Stefan Beller <sbeller@google.com>