config: prepare copy/rename section function for arbitrary repositories
The base function 'git_config_copy_or_rename_section_in_file' deals with
arbitrary repositories gracefully already, except in the (common) corner
case of 'config_filename == NULL', which asks the function to additionally
lookup the path of the default config in the main repository.

We could pass in a repository to 'git_config_copy_or_rename_section_in_file'
that would use the repository specific lookup function 'repo_git_path'
for the path, but this is not the best way to approach the problem as it
seems like a layering violation.

Instead move the code to handle the corner case of a NULL config_filename
out of the base function. It looks like code is duplicated, but this will
change in the next patch, when we use an arbitrary repository in 2 out of
the 3 wrapper functions.

Signed-off-by: Stefan Beller <sbeller@google.com>