submodule.c: check submodule remotes including broken remote refs
Currently when checking if a submodule needs pushing, we only inspect
if there is at least one remote and then we'll exec a new process to run
a more detailed test. The check if there is a remote is done via iterating
over refs prefixed by "refs/remotes/" in the submodule using
'for_each_remote_ref_submodule' with a callback 'has_remote' that just
returns true for any ref.

However if a broken ref is given to the the ref iterating machinery aborts
with an error, such that the check
    if (for_each_remote_ref_submodule(path, has_remote, NULL) > 0) {
fails as the return is -1.

For the original goal of determining if there are any remotes it is ok
to allow for broken refs in refs/remotes, such that "any ref [including
broken refs] under refs/remotes indicate the existence of a remote".

By allowing for broken refs, via the refs_for_each_fullref_in function,
we do not need to add the objects of the submodule as an alternative,
such that we can lose the two calls to add_submodule_odb.

Signed-off-by: Stefan Beller <sbeller@google.com>