submodule: use submodule repos for object lookup
This converts the 'show_submodule_header' function to use
the repository API properly, such that the submodule objects
are not added to the main object store.

When using the inline diff of submodules, the object store is not
polluted with objects from the submodules. When using the "log"
summary, we still have to pollute the object store, as the
revision walking is not converted yet.

Signed-off-by: Stefan Beller <sbeller@google.com>