pack-bitmap.c: migrate its global variable into the object store
There is only one global variable in the bitmap code, which
is a struct bitmap_index.

Migrate the variable into the repositories raw object store,
where we load it from in each function needing access.

As we store a pointer in the object store we have to make
sure to allocate memory for the bitmap index before using it.
There are 2 entry points into the bitbmap code,
prepare_bitmap_git() as well as prepare_bitmap_walk().
Both of them check if they need to allocate memory and do so
before proceeding.

This patch serves two purposes, firstly it is a preparatory
step towards migrating the bitmap code into the object store
completely (further patches will need to pass through the
repository object or object store such that we can operate
on other repositories than the_repository); secondly
this patch allows for memory cleanup when free'ing the
object store. Currently the object store lacks free'ing
'struct packed_git*' pack files as the bitmap code would
have dangling pointers to those pack files. By migrating
the bitmaps to the object store a follow up patch can
free the packfiles and make sure there are no dangling
pointers left in the bitmap code.

Signed-off-by: Stefan Beller <sbeller@google.com>