refs.c: introduce for_each_replace_repo_ref to take a repository argument
The function 'for_each_replace_repo_ref' will be the repository specific
version of 'for_each_replace_ref'.

Instead of just adding the new function, replace 'for_each_replace_ref'
by a shallow wrapper to the new function as well. This helps chained
history blaming, when digging for the origin of this function.

Signed-off-by: Stefan Beller <sbeller@google.com>