xdiff/xpatience: duplicate insert_records and hardcode pass argument
The function 'insert_records' is used both for inserting records from the
one side as well as checking for duplicates from the other side, which
is distinguished by the 'pass' argument.

Duplicate the function to a new function 'match_record', remove the 'pass'
argument and hardcode it in each function.

A follow up patch will remove the dead code.

Signed-off-by: Stefan Beller <sbeller@google.com>